backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4

openai==1.24.1
langchain==0.1.20
langchain-core==0.1.38
langchain-openai==0.1.7
langchain-anthropic==0.1.4

browser-use
typing-extensions>=4.5.0
